+++ src/main/java/com/gmail/val59000mc/game/GameManager.java
@@ -329,7 +329,6 @@ public class GameManager{
 		listeners.add(new BlockListener(this));
 		listeners.add(new WorldListener());
 		listeners.add(new PlayerMovementListener(playerManager));
-		listeners.add(new EntityDamageListener(this));
 		listeners.add(new PlayerHungerGainListener(playerManager));
 		for(Listener listener : listeners){
 			Bukkit.getServer().getPluginManager().registerEvents(listener, UhcCore.getPlugin());